requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
alembic==1.16.5
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.2
attrs==25.4.0
billiard==4.2.2
blinker==1.9.0
cachetools==6.2.1
celery==5.5.3
celery-redbeat==2.3.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.117.1
Flask==3.1.2
Flask-Admin==2.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
frozenlist==1.8.0
git-filter-repo==2.47.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
greenlet==3.2.4
h11==0.16.0
http_ece==1.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
load-dotenv==0.1.0
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
numpy==2.3.3
oauthlib==3.3.1
packaging==25.0
pandas==2.3.2
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.1.3
psycopg2-binary==2.9.10
py-vapid==1.9.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywebpush==2.1.1
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
stripe==12.5.1
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.3
wheel==0.45.1
WTForms==3.2.1
yarl==1.22.0
